refactor: Update topProducts import paths to use hooks folder
