+++ src/components/topProducts.tsx
@@ -1,4 +1,4 @@
-import { COLORS } from '@/utils/colors'
+import { COLORS } from '@/hooks/colors'
 
 export interface ProductSales {
   productName: string
